mini_google_search/requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
PyPDF2
pdfminer.six
# Optional:
# redis
# numpy
# sentence-transformers
# faiss-cpu
